+++ CHANGELOG.md
@@ -1,2 +1,2 @@
 # Evolution Log
-- 2026-07-13T16:31:41.476Z: Pulse 96410 - Integrity Stable.
+- 2026-07-13T16:32:34.902Z: Pulse 96420 - Integrity Stable.